Change sponge's tests as integration tests
This also address the review's comments and add few minor code refactors.
